Add e2e tests for GitHub

+++ git/internal/e2e/run.sh
@@ -8,12 +8,17 @@ DIR="$(cd "$(dirname "$0")" && pwd)"
 cd "${PROJECT_DIR}/git/libgit2" && make libgit2
 LIBGIT2_BUILD_DIR=${PROJECT_DIR}/git/libgit2/build . "${PROJECT_DIR}/git/libgit2/libgit2-vars.env"
 
-source "${DIR}/setup_gitlab.sh"
+if [[ "${GO_TESTS}" = "" ]] || [[ "${GO_TESTS}" = "-run TestGitLabE2E" ]]; then
+    source "${DIR}/setup_gitlab.sh"
+fi
 
 cd "${DIR}"
 CGO_LDFLAGS=$(PKG_CONFIG_PATH="${PKG_CONFIG_PATH}" pkg-config --libs --static --cflags libgit2)
-PKG_CONFIG_PATH="${PKG_CONFIG_PATH}" CGO_LDFLAGS="${CGO_LDFLAGS}" CGO_ENABLED=1 go test -v -tags 'netgo,osusergo,static_build,e2e' -race ./...
+PKG_CONFIG_PATH="${PKG_CONFIG_PATH}" CGO_LDFLAGS="${CGO_LDFLAGS}" CGO_ENABLED=1 \
+    go test -v -tags 'netgo,osusergo,static_build,e2e' -race ${GO_TESTS} ./...
 
 # cleanup
-docker kill gitlab && docker rm gitlab
+if [[ "${GO_TESTS}" = "" ]] || [[ "${GO_TESTS}" = "-run TestGitLabE2E" ]]; then
+    docker kill gitlab && docker rm gitlab
+fi
